big ass commit at 1am
- refactored a lot of internal code
- separated a lot of classes into their own files
- nothing broke
- added editor timeline and preview display
- :o works flawlessly
