+++ .gitignore
@@ -1,4 +1,5 @@
 /__pycache__
+/*/__pycache__
 /charts/*/
 /.vscode/
 *.so
